requirements.txt
apache-airflow
apache-airflow-providers-amazon
apache-airflow-providers-apache-spark
apache-airflow-providers-ftp
apache-airflow-providers-http
apache-airflow-providers-imap
apache-airflow-providers-sqlite
apache-airflow-providers-mongo
backoff
pyspark
requests
feedparser
redis
Unidecode
kafka-python
minio
boto3
emails
pandas
cchardet
papermill
plotly
plotly_express
kaleido
ipykernel
nbconvert
openpyxl
pymongo